backend/requirements.txt
aider-chat==0.12.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.9.3
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
ConfigArgParse==1.7
dataclasses-json==0.5.14
diskcache==5.6.1
exceptiongroup==1.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.283
langsmith==0.0.33
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.5
numpy==1.24.3
openai==0.27.6
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.38
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.15.1
PyPDF2==3.0.1
pyrsistent==0.19.3
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.30.0
rich==13.3.5
scipy==1.10.1
smmap==5.0.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
SQLAlchemy==2.0.20
tenacity==8.2.3
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.7
yarl==1.9.2
